+++ src/object_cmd.cpp
@@ -511,7 +511,7 @@ static void ReallyClearObjectTile(Object *o)
 {
 	Object::DecTypeCount(o->type);
 	for (TileIndex tile_cur : o->location) {
-		DeleteNewGRFInspectWindow(GSF_OBJECTS, tile_cur);
+		DeleteNewGRFInspectWindow(GSF_OBJECTS, static_cast<uint32_t>(tile_cur));
 
 		MakeWaterKeepingClass(tile_cur, GetTileOwner(tile_cur));
 	}
